refactored redux store. TODO: refactor react router
