+++ server/client/.eslintrc
@@ -33,6 +33,7 @@
         ],
         "consistent-return": [
             "off"
-        ]
+        ],
+        "no-shadow": ["error", { "hoist": "all" }]
     }
 }
