+++ server/client/src/index.js
@@ -3,21 +3,9 @@ import 'materialize-css/dist/css/materialize.min.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
-import reduxThunk from 'redux-thunk';
 
+import store from './redux/store';
 import App from './components/App';
-import reducers from './reducers';
-
-const initialState = {
-	auth: { data: false, loading: true },
-};
-
-const store = createStore(
-	reducers,
-	initialState,
-	applyMiddleware(reduxThunk),
-);
 
 ReactDOM.render(
 	<Provider store={store}>
