+++ server/client/src/redux/actions/fetchUser.js
@@ -1,0 +1,9 @@
+import axios from 'axios';
+import { FETCH_USER } from '../types';
+
+const fetchUser = () => async (dispatch) => {
+	const res = await axios.get('/api/current_user');
+	dispatch({ type: FETCH_USER, payload: res.data });
+};
+
+export default fetchUser;
