+++ server/client/src/redux/actions/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export { default as fetchUser } from './fetchUser';
