+++ server/client/src/redux/reducers/authReducer.js
@@ -1,4 +1,4 @@
-import { FETCH_USER } from '../actions/types';
+import { FETCH_USER } from '../types';
 
 export default (state = null, action) => {
 	switch (action.type) {
