+++ server/client/src/redux/store/index.js
@@ -1,0 +1,13 @@
+import { createStore, applyMiddleware } from 'redux';
+import reduxThunk from 'redux-thunk';
+
+import initialState from './initialState';
+import reducers from '../reducers';
+
+const store = createStore(
+	reducers,
+	initialState,
+	applyMiddleware(reduxThunk),
+);
+
+export default store;
