+++ server/client/src/redux/store/initialState.js
@@ -1,0 +1,5 @@
+const initialState = {
+	auth: { data: false, loading: true },
+};
+
+export default initialState;
