+++ server/client/src/redux/types.js
